backend/requirements.txt
numpy == 1.23.5
python-dotenv == 1.0.0
pandas == 1.5.3
pymorphy2 == 0.8
joblib == 1.1.1
pydantic == 1.10.9
scikit-learn == 1.2.2
scipy == 1.10.1
APScheduler == 3.10.1
fastapi == 0.97.0
beautifulsoup4 == 4.12.2
selenium == 4.10.0
webdriver-manager == 3.8.6
uvicorn == 0.22.0
requests == 2.28.1
pyspellchecker == 0.7.2